requirements.txt
rembg
easygui
Pillow
onnxruntime
Flask
